LayoutTests/imported/w3c/web-platform-tests/html/webappapis/dynamic-markup-insertion/opening-the-input-stream/url-fragment.window-expected.txt


PASS document.open() and document's URL containing a fragment (entry is not relevant)
PASS document.open() and document's URL containing a fragment (entry is relevant)
